refactor(web): finish constants and feedback migrations
Point every caller at the domain constants and feedback modules, then delete
the two compatibility barrels. The narrower imports make ownership explicit
and stop temporary migration surfaces from becoming permanent APIs.

Refresh the Biome schema. Use an ESM-native Vitest config so both tools run
without stale configuration warnings.
