+++ apps/web/src/app/(main)/epics/page.tsx
@@ -4,12 +4,12 @@ import { useRouter, useSearchParams } from 'next/navigation';
 import { Suspense, useCallback, useEffect, useMemo, useState } from 'react';
 import { EpicList } from '@/components/epics';
 import { PageHeader } from '@/components/layout/page-header';
+import { ErrorState } from '@/components/ui/empty-state';
 import { ChevronDown, Search, X } from '@/components/ui/icons';
 import { LoadingState } from '@/components/ui/spinner';
 import { FilterChip } from '@/components/ui/toggle';
-import { ErrorState } from '@/components/ui/tooltip';
 import type { EpicStatus } from '@/lib/api';
-import { EPIC_STATUS_CONFIG } from '@/lib/constants';
+import { EPIC_STATUS_CONFIG } from '@/lib/constants/epics';
 import { useEpics, useProjects } from '@/lib/hooks';
 import { useProjectFilters } from '@/lib/project-context';
 import { readStorage, writeStorage } from '@/lib/storage';
