+++ apps/web/src/components/tasks/kanban-board.tsx
@@ -13,7 +13,7 @@ import {
   Zap,
 } from '@/components/ui/icons';
 import type { TaskStatus, TaskSummary } from '@/lib/api';
-import { TASK_STATUS_CONFIG, TASK_STATUSES, type TaskStatusType } from '@/lib/constants';
+import { TASK_STATUS_CONFIG, TASK_STATUSES, type TaskStatusType } from '@/lib/constants/tasks';
 import { useProjectContext, useProjectFilter } from '@/lib/project-context';
 import { TaskCard, TaskCardSkeleton } from './task-card';
