+++ apps/web/src/components/tasks/task-card.tsx
@@ -3,7 +3,7 @@
 import { memo } from 'react';
 import { AlertTriangle, Calendar, Clock, LightBulb, Pause, Zap } from '@/components/ui/icons';
 import type { TaskPriority, TaskStatus, TaskSummary } from '@/lib/api';
-import { TASK_STATUS_CONFIG } from '@/lib/constants';
+import { TASK_STATUS_CONFIG } from '@/lib/constants/tasks';
 
 interface TaskCardProps {
   task: TaskSummary;
