+++ apps/web/src/components/tasks/task-list-mobile.tsx
@@ -4,7 +4,7 @@ import { AnimatePresence, motion } from 'motion/react';
 import { memo, useMemo, useState } from 'react';
 import { ArrowDownAZ, Calendar, Clock, Zap } from '@/components/ui/icons';
 import type { TaskStatus, TaskSummary } from '@/lib/api';
-import { TASK_STATUS_CONFIG, type TaskStatusType } from '@/lib/constants';
+import { TASK_STATUS_CONFIG, type TaskStatusType } from '@/lib/constants/tasks';
 import { useProjectContext, useProjectFilter } from '@/lib/project-context';
 import { TaskCard } from './task-card';
