--- apps/web/src/lib/constants.ts
@@ -1,13 +1,0 @@
-// =============================================================================
-// SilkCircuit Design System Constants
-// =============================================================================
-//
-// This file re-exports from domain-specific modules for backwards compatibility.
-// For new imports, prefer importing directly from the specific module:
-//
-//   import { TASK_STATUS_CONFIG } from '@/lib/constants/tasks';
-//   import { ENTITY_COLORS } from '@/lib/constants/entities';
-//
-// =============================================================================
-
-export * from './constants/index';
